backend/requirements.txt
Flask
Flask-Cors
pymongo
gunicorn
numpy
opencv-python
pycryptodome
requests
web3
google-generativeai
python-dotenv
